Moved actions to their own file and modularized it
